refactor: keep spend export modules within budget
